# api/openapi/request-to-pay-service.yaml
openapi: 3.0.3
info:
  title: Request to Pay Service API
  version: 1.0.0
  description: >-
    TPP-facing request-to-pay API of svc-pay-request-to-pay. Paths are the
    monolith's (/open-finance/v1/par, /open-finance/v1/payment-consents/...).
    Every operation requires a DPoP-bound access token (RFC 9449): the
    Authorization header uses the DPoP scheme, a DPoP proof with a unique jti
    accompanies each request, the token's cnf.jkt matches the proof key and its
    aud contains svc-pay-request-to-pay. A Bearer token is answered with 401
    and a WWW-Authenticate DPoP challenge.
servers:
  - url: https://api.example.com/open-finance/v1
paths:
  /par:
    post:
      summary: Create request to pay
      operationId: createRequestToPay
      security:
        - dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/FinancialId'
        - $ref: '#/components/parameters/Dpop'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestToPay'
      responses:
        '201':
          description: >-
            Request created. A retry with the same X-Idempotency-Key and payload
            returns the first request with X-Idempotent-Replay true.
          headers:
            Location:
              schema:
                type: string
            X-Idempotent-Replay:
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestToPayCreated'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/DpopChallenge'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /payment-consents/{consentId}:
    get:
      summary: Get request to pay status
      operationId: getRequestToPay
      security:
        - dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/FinancialId'
        - $ref: '#/components/parameters/Dpop'
        - $ref: '#/components/parameters/ConsentId'
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Request status
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestToPayStatus'
        '304':
          description: Not modified (ETag matches If-None-Match)
        '401':
          $ref: '#/components/responses/DpopChallenge'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /payment-consents/{consentId}/accept:
    post:
      summary: Accept request to pay
      operationId: acceptRequestToPay
      security:
        - dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/FinancialId'
        - $ref: '#/components/parameters/Dpop'
        - $ref: '#/components/parameters/ConsentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Decision'
      responses:
        '201':
          description: Request accepted (status Consumed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestToPayStatus'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/DpopChallenge'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /payment-consents/{consentId}/reject:
    post:
      summary: Reject request to pay
      operationId: rejectRequestToPay
      security:
        - dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/FinancialId'
        - $ref: '#/components/parameters/Dpop'
        - $ref: '#/components/parameters/ConsentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Decision'
      responses:
        '200':
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestToPayStatus'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/DpopChallenge'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    dpopAuth:
      type: http
      scheme: DPoP
      description: >-
        Authorization: DPoP <access token> from the fintechbankx realm, with
        cnf.jkt bound to the key of the DPoP proof header; aud must contain
        svc-pay-request-to-pay.
  parameters:
    InteractionId:
      name: x-fapi-interaction-id
      in: header
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9._:-]{1,128}$'
    IdempotencyKey:
      name: x-idempotency-key
      in: header
      required: true
      schema:
        type: string
        maxLength: 128
    FinancialId:
      name: x-fapi-financial-id
      in: header
      required: false
      description: May repeat the calling TPP's client id; any other value is 403.
      schema:
        type: string
    Dpop:
      name: DPoP
      in: header
      required: true
      description: >-
        DPoP proof JWT (RFC 9449), typ dpop+jwt, ES256 or PS256, with htm, htu,
        iat and a jti that has not been used before.
      schema:
        type: string
    ConsentId:
      name: consentId
      in: path
      required: true
      description: Id of the pay request (returned as Data.ConsentId on create).
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    DpopChallenge:
      description: Missing or invalid DPoP-bound token or proof, or a non-DPoP scheme
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: DPoP error="invalid_dpop_proof", algs="ES256 PS256"
  schemas:
    CreateRequestToPay:
      type: object
      required: [Data]
      properties:
        Data:
          type: object
          required: [PsuId, CreditorName, InstructedAmount]
          properties:
            PsuId:
              type: string
              maxLength: 128
            CreditorName:
              type: string
              maxLength: 140
            InstructedAmount:
              type: object
              required: [Amount, Currency]
              properties:
                Amount:
                  type: string
                  pattern: '^[0-9]{1,15}(\.[0-9]{1,4})?$'
                  description: No more decimals than the currency's minor unit.
                Currency:
                  type: string
                  pattern: '^[A-Za-z]{3}$'
                  description: ISO 4217 code.
    Decision:
      type: object
      properties:
        paymentId:
          type: string
          description: Required to accept.
        reason:
          type: string
    RequestToPayCreated:
      type: object
      properties:
        Data:
          type: object
          properties:
            ConsentId:
              type: string
            Status:
              $ref: '#/components/schemas/Status'
        Links:
          $ref: '#/components/schemas/Links'
    RequestToPayStatus:
      type: object
      properties:
        Data:
          type: object
          properties:
            ConsentId:
              type: string
            Status:
              $ref: '#/components/schemas/Status'
            PaymentId:
              type: string
        Links:
          $ref: '#/components/schemas/Links'
    Status:
      type: string
      enum: [AwaitingAuthorisation, Consumed, Rejected]
    Links:
      type: object
      properties:
        Self:
          type: string
    Error:
      type: object
      properties:
        code:
          type: string
          enum: [INVALID_REQUEST, FORBIDDEN, NOT_FOUND, REQUEST_FINALIZED, IDEMPOTENCY_KEY_REUSED,
                 CONCURRENT_UPDATE, DPOP_VALIDATION_FAILED, INTERNAL_ERROR]
        message:
          type: string
        interactionId:
          type: string
        timestamp:
          type: string
          format: date-time