# open-finance-infrastructure/src/main/java/com/enterprise/openfinance/requesttopay/infrastructure/rest/PayRequestController.java
package com.enterprise.openfinance.requesttopay.infrastructure.rest;

import com.enterprise.openfinance.requesttopay.domain.port.in.PayRequestUseCase;
import com.enterprise.openfinance.requesttopay.domain.query.GetPayRequestStatusQuery;
import com.enterprise.openfinance.requesttopay.infrastructure.rest.dto.PayRequestDecisionRequest;
import com.enterprise.openfinance.requesttopay.infrastructure.rest.dto.PayRequestRequest;
import com.enterprise.openfinance.requesttopay.infrastructure.rest.dto.PayRequestResponse;
import com.enterprise.openfinance.requesttopay.infrastructure.rest.dto.PayRequestStatusResponse;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;
import org.springframework.http.CacheControl;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import java.nio.charset.StandardCharsets;
import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.util.Base64;
import java.util.concurrent.TimeUnit;

@RestController
@Validated
@RequestMapping("/open-finance/v1")
public class PayRequestController {

    /** TPP-facing paths, kept as in the monolith; DPoP is required on all of them. */
    static final String SELF_PREFIX = "/open-finance/v1/payment-consents/";

    /** x-fapi-interaction-id becomes the event correlationId: a single safe token, at most 128 characters. */
    static final String INTERACTION_ID = "^[A-Za-z0-9._:-]{1,128}$";

    private final PayRequestUseCase useCase;

    public PayRequestController(PayRequestUseCase useCase) {
        this.useCase = useCase;
    }

    @PostMapping("/par")
    @DPoPSecured
    @FAPISecured
    public ResponseEntity<PayRequestResponse> createPayRequest(
            @RequestHeader("X-FAPI-Interaction-ID") @NotBlank @Pattern(regexp = INTERACTION_ID) String interactionId,
            @RequestHeader("X-Idempotency-Key") @NotBlank @Size(max = 128) String idempotencyKey,
            @RequestHeader(value = "x-fapi-financial-id", required = false) String financialId,
            @RequestBody @Valid PayRequestRequest request
    ) {
        // Idempotency: a retry with the same key and payload returns the first result
        // (same consent id, no new events); a different payload is 409.
        String tppId = TppIdentity.resolve(financialId);
        var result = useCase.createPayRequest(request.toCommand(tppId, interactionId, idempotencyKey));
        String self = SELF_PREFIX + result.request().consentId();

        return ResponseEntity.created(java.net.URI.create(self))
                .cacheControl(CacheControl.maxAge(0, TimeUnit.SECONDS).noStore())
                .header("X-FAPI-Interaction-ID", interactionId)
                .header("X-OF-Cache", result.cacheHit() ? "HIT" : "MISS")
                .header("X-Idempotent-Replay", Boolean.toString(result.idempotentReplay()))
                .body(PayRequestResponse.from(result, self));
    }

    @GetMapping("/payment-consents/{consentId}")
    @DPoPSecured
    @FAPISecured
    public ResponseEntity<PayRequestStatusResponse> getPayRequestStatus(
            @RequestHeader("X-FAPI-Interaction-ID") @NotBlank @Pattern(regexp = INTERACTION_ID) String interactionId,
            @RequestHeader(value = "x-fapi-financial-id", required = false) String financialId,
            @PathVariable("consentId") @NotBlank String consentId,
            @RequestHeader(value = "If-None-Match", required = false) String ifNoneMatch
    ) {
        String tppId = TppIdentity.resolve(financialId);

        var result = useCase.getPayRequestStatus(new GetPayRequestStatusQuery(consentId, tppId, interactionId));
        String self = SELF_PREFIX + consentId;
        PayRequestStatusResponse response = PayRequestStatusResponse.from(result, self);
        String etag = generateEtag(response.data().toString());

        if (ifNoneMatch != null && ifNoneMatch.equals(etag)) {
            return ResponseEntity.status(HttpStatus.NOT_MODIFIED)
                    .cacheControl(CacheControl.maxAge(0, TimeUnit.SECONDS).noStore())
                    .header("X-FAPI-Interaction-ID", interactionId)
                    .eTag(etag)
                    .build();
        }

        return ResponseEntity.ok()
                .cacheControl(CacheControl.maxAge(0, TimeUnit.SECONDS).noStore())
                .header("X-FAPI-Interaction-ID", interactionId)
                .header("X-OF-Cache", result.cacheHit() ? "HIT" : "MISS")
                .eTag(etag)
                .body(response);
    }

    @PostMapping("/payment-consents/{consentId}/accept")
    @DPoPSecured
    @FAPISecured
    public ResponseEntity<PayRequestStatusResponse> acceptPayRequest(
            @RequestHeader("X-FAPI-Interaction-ID") @NotBlank @Pattern(regexp = INTERACTION_ID) String interactionId,
            @RequestHeader(value = "x-fapi-financial-id", required = false) String financialId,
            @PathVariable("consentId") @NotBlank String consentId,
            @RequestBody PayRequestDecisionRequest decision
    ) {
        String tppId = TppIdentity.resolve(financialId);

        String paymentId = decision == null ? null : decision.paymentId();
        var result = useCase.acceptPayRequest(consentId, tppId, paymentId, interactionId);
        String self = SELF_PREFIX + consentId;

        return ResponseEntity.status(HttpStatus.CREATED)
                .cacheControl(CacheControl.maxAge(0, TimeUnit.SECONDS).noStore())
                .header("X-FAPI-Interaction-ID", interactionId)
                .body(PayRequestStatusResponse.from(result, self));
    }

    @PostMapping("/payment-consents/{consentId}/reject")
    @DPoPSecured
    @FAPISecured
    public ResponseEntity<PayRequestStatusResponse> rejectPayRequest(
            @RequestHeader("X-FAPI-Interaction-ID") @NotBlank @Pattern(regexp = INTERACTION_ID) String interactionId,
            @RequestHeader(value = "x-fapi-financial-id", required = false) String financialId,
            @PathVariable("consentId") @NotBlank String consentId,
            @RequestBody PayRequestDecisionRequest decision
    ) {
        String tppId = TppIdentity.resolve(financialId);

        var result = useCase.rejectPayRequest(consentId, tppId, interactionId);
        String self = SELF_PREFIX + consentId;

        return ResponseEntity.ok()
                .cacheControl(CacheControl.maxAge(0, TimeUnit.SECONDS).noStore())
                .header("X-FAPI-Interaction-ID", interactionId)
                .body(PayRequestStatusResponse.from(result, self));
    }

    private static String generateEtag(String payload) {
        try {
            MessageDigest digest = MessageDigest.getInstance("SHA-256");
            byte[] hash = digest.digest(payload.getBytes(StandardCharsets.UTF_8));
            return '"' + Base64.getUrlEncoder().withoutPadding().encodeToString(hash) + '"';
        } catch (NoSuchAlgorithmException exception) {
            throw new IllegalStateException("Unable to generate ETag", exception);
        }
    }
}